requirements.txt
GitPython
openai
langchain
pydantic
python-dotenv
langchain-core
radon